Fix memory leak
If new BufferedReader throws an exception, an instance of FileReader
would be open upon destruction by garbage collector because it wouldn't
be assigned to any variable and the finally block would not close it
because reader would be null, therefore causing a leak.

To address the issue, try-with-resources functionality of Java 7 is
used. This get the warnings right, and makes the code simpler, also
saving a finally block.